+++ src/main/java/org/mybatis/dynamic/sql/insert/render/DefaultInsertStatementProvider.java
@@ -21,24 +21,11 @@ import org.jetbrains.annotations.NotNull;
 
 public class DefaultInsertStatementProvider<T> implements InsertStatementProvider<T> {
     private final String insertStatement;
-    // need to keep both row and record for now so we don't break
-    // old code. The MyBatis reflection utilities don't handle
-    // the case where the attribute name is different from the getter.
-    //
-    // MyBatis Generator version 1.4.1 (March 8, 2022) changed to use "row" instead of "record".
-    // Target March 2023 for removing "record" from MyBatis Dynamic SQL.
-    private final T record;
     private final T row;
 
     private DefaultInsertStatementProvider(Builder<T> builder) {
         insertStatement = Objects.requireNonNull(builder.insertStatement);
         row = Objects.requireNonNull(builder.row);
-        record = row;
-    }
-
-    @Override
-    public T getRecord() {
-        return record;
     }
 
     @Override